+++ src/legends_ultimate_captions/__init__.py
@@ -1,4 +1,0 @@
-"""Legends Ultimate Captions package."""
-
-__version__ = "0.1.0"
-